+++ src/rfc7591/mod.rs
@@ -3,3 +3,4 @@
 //! Refs: <https://datatracker.ietf.org/doc/html/rfc7591>
 
 pub mod register;
+pub mod source;